The services which interact with the typescript models also have to be wirtten in typescript so that the type safety of themodels is protected
